pyproject.toml
[project]
name = "test-engry"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic-settings>=2.6.1",
    "python-dotenv>=1.0.1",
    "python-telegram-bot==21.8",
]
[tool.pyright]
venvPath = "."
venv = ".venv"

[tool.isort]
profile = "black"

[tool.autoflake]
check = true

[tool.ruff.lint]
extend-select = ["I"]

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"
